fix a reversed size_t subtraction, and make narrowing explicit
game_features.cpp had the operands backwards:

    const auto initialSize = features.size();   [before erase]
    features.erase(...);                        [shrinks]
    const int removed = features.size() - initialSize;

The subtraction is size_t, so removing anything underflows and the result
narrows to a negative int. unregisterGameFeatures() therefore returned -N
instead of N, and gamefeaturesproxy passes that straight out to the plugin
API. The `if (removed)` guard still fired, so the side effect was right and
only the count was wrong. Swapped the operands and cast.

C5054: two files compared a value against the wrong enum. Profile has a
QDialogButtonBox::StandardButton from a QuestionBoxMemory query but compared
it to QMessageBox::Yes/No; spawn.cpp has a QMessageBox::StandardButton from
confirmStartSteam but compared it to QDialogButtonBox::Yes/No. The numeric
values happen to coincide today, which is why nobody noticed. Each now
compares against its own enum.

C4267: the remaining size_t to int narrowing is all container sizes and loop
indices fed to Qt APIs that take int, bounded in practice by the mod count.
Made explicit with static_cast rather than left implicit.

C4505: removed two dead static functions, ByName in modinforegular.cpp and
SupportOptimizedFind in directoryentry.cpp, the latter a Windows 7 era
version check that nothing calls any more.

C4267, C5054 and C4505 are now at zero. 0 errors.

+++ src/disableproxyplugindialog.cpp
@@ -13,7 +13,7 @@ DisableProxyPluginDialog::DisableProxyPluginDialog(
 
   ui->topLabel->setText(QObject::tr("Disabling the '%1' plugin will prevent the "
                                     "following %2 plugin(s) from working:",
-                                    "", required.size())
+                                    "", static_cast<int>(required.size()))
                             .arg(proxyPlugin->localizedName())
                             .arg(required.size()));
 
@@ -21,7 +21,7 @@ DisableProxyPluginDialog::DisableProxyPluginDialog(
   connect(ui->yesBtn, &QPushButton::clicked, this, &QDialog::accept);
 
   ui->requiredPlugins->setSelectionMode(QAbstractItemView::NoSelection);
-  ui->requiredPlugins->setRowCount(required.size());
+  ui->requiredPlugins->setRowCount(static_cast<int>(required.size()));
   for (int i = 0; i < required.size(); ++i) {
     ui->requiredPlugins->setItem(i, 0,
                                  new QTableWidgetItem(required[i]->localizedName()));
